// server/src/modules/payments/dto/verify-payment.dto.ts
import { IsNotEmpty, IsString } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class VerifyPaymentDto {
  @ApiProperty({
    example: 'LST-89421056',
    description: 'System order identifier',
  })
  @IsString()
  @IsNotEmpty({ message: 'Order ID is required.' })
  orderId: string;

  @ApiProperty({
    example: 'order_EkWnAWifjBnga6',
    description: 'Razorpay session order ID returned by create-intent',
  })
  @IsString()
  @IsNotEmpty({ message: 'Razorpay order ID is required.' })
  razorpayOrderId: string;

  @ApiProperty({
    example: 'pay_29QQoUBcxrhLEZ',
    description: 'Razorpay payment ID generated upon customer authorization',
  })
  @IsString()
  @IsNotEmpty({ message: 'Razorpay payment ID is required.' })
  razorpayPaymentId: string;

  @ApiProperty({
    example: '9ef5262a4e6ff202b291ae5d32520773be1af0e493cb05851881c787323082da',
    description: 'Cryptographic HMAC-SHA256 signature generated by Razorpay',
  })
  @IsString()
  @IsNotEmpty({ message: 'Razorpay signature is required.' })
  razorpaySignature: string;
}